requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.3
dataclasses==0.8
discord.py==1.7.3
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
pkg-resources==0.0.0
prettytable==2.4.0
python-dotenv==0.19.2
typing-extensions==4.0.1
wcwidth==0.2.5
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
